+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The main reasoning behind this choice is the software is not built to be extende
 
 - Prevent empty comments being submitted to user events history.
 - Correctly render HTML tags on investigation pages.
+- Rename subject to user for Reports.
 
 ## Version 2 (2022-06-27)